separated all concerns to different files
